# contactapi/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "STIS Contact API",
    "description": "API ini menyediakan layanan untuk mengelola dan mengakses informasi kontak yang dimiliki oleh mahasiswa dan dosen di Politeknik Statistika STIS. Layanan ini dirancang untuk mempermudah pencarian informasi kontak dengan menyediakan berbagai endpoint untuk menambah, memperbarui, dan menghapus kontak dari mahasiswa dan dosen. API ini juga mendukung pengelompokan kontak berdasarkan organisasi mahasiswa atau mata kuliah yang diampu oleh dosen serta menawarkan fitur pencarian yang fleksibel. \n\nFitur utama meliputi: \n- Manajemen kontak mahasiswa dan dosen \n- Pengelompokan kontak berdasarkan organisasi atau jurusan \n- Dukungan endpoint untuk CRUD (Create, Read, Update, Delete) pada data kontak \n- Fitur pencarian kontak berdasarkan organisasi atau mata kuliah yang diampu oleh dosen\n\n",
    "contact": {
      "name": "Dodi Firmansyah",
      "url": "https://github.com/dodfrm",
      "email": "222212572@stis.ac.id"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/organizations/{id}": {
      "get": {
        "tags": ["Organization"],
        "summary": "Get organization by ID",
        "operationId": "getOrganizationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Organization not found"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Organization found",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Organization"],
        "summary": "Update an organization",
        "operationId": "updateOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "organizationName": "BULSTIK"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Organization not found"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Organization updated successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Organization"],
        "summary": "Delete an organization",
        "operationId": "deleteOrganization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Organization not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Organization not found"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Organization deleted successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/organizations": {
      "get": {
        "tags": ["Organization"],
        "summary": "Get all organizations",
        "operationId": "getAllOrganizations",
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "200": {
            "description": "List of all organizations",
            "content": {
              "application/hal+json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Organization"],
        "summary": "Create a new organization",
        "operationId": "createOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "string",
                "example": {
                  "organizationName": "BULSTIK"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": {
                    "message": "Unauthorized"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          },
          "201": {
            "description": "Organization created successfully",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Subject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subjectName": {
            "type": "string"
          }
        }
      },
      "ContactSubject": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "subject": {
            "$ref": "#/components/schemas/Subject"
          }
        }
      },
      "ContactOrganization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "kelas": {
            "type": "string"
          },
          "jabatan": {
            "type": "string",
            "enum": ["KETUA", "WAKIL_KETUA", "BENDAHARA", "PJ_MATKUL"]
          },
          "periodeJabatan": {
            "type": "string"
          }
        }
      },
...
      "ContactDto": {
        "required": ["fullName", "phone"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fullName": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "contactType": {
            "type": "string",
            "enum": ["MAHASISWA", "DOSEN"]
          },
          "contactOrganizations": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactOrganization"
            }
          },
          "contactSubjects": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactSubject"
            }
          }
        }
      },
...
      "OrganizationDto": {
        "required": ["organizationName"],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "organizationName": {
            "type": "string"
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContactDto"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Input token sebagai 'Bearer \u003Ctoken\u003E'",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}